requirements.txt
beautifulsoup4==4.12.3
Markdown==3.5.2
PyYAML==6.0.1
flask==3.0.3